move get_name to get_name_from_filename and add unit tests
