+++ package.json
@@ -24,7 +24,7 @@
   "types": "types/runtime",
   "scripts": {
     "test": "mocha --opts mocha.opts",
-    "test:unit": "mocha --require sucrase/register --recursive ./**/__test__.ts",
+    "test:unit": "mocha --require sucrase/register --recursive src/**/__test__.ts",
     "quicktest": "mocha --opts mocha.opts",
     "precoverage": "c8 mocha --opts mocha.coverage.opts",
     "coverage": "c8 report --reporter=text-lcov > coverage.lcov && c8 report --reporter=html",
